Update uiSize for pixel density handling

+++ src/main/java/heronarts/glx/ui/UI.java
@@ -720,11 +720,11 @@ public class UI {
     return this.lx.getContentScaleY();
   }
 
-  public int getWidth() {
+  public float getWidth() {
     return this.lx.getUIWidth();
   }
 
-  public int getHeight() {
+  public float getHeight() {
     return this.lx.getUIHeight();
   }
